+++ CMakeLists.txt
@@ -15,7 +15,7 @@ cmake_policy(SET CMP0015 NEW)
 
 set(GUI text.cpp)
 set(UTILS util.cpp)
-set(MATH quaternion.cpp vector.cpp matrix.cpp ray.cpp)
+set(MATH quaternion.cpp vector.cpp matrix.cpp rayCaster.cpp)
 set(RENDER lineRenderer.cpp particleEmitter.cpp camera.cpp light.cpp material.cpp mesh.cpp meshData.cpp model.cpp node.cpp quad.cpp box.cpp root.cpp shader.cpp texture.cpp)
 set(ANIM animationController.cpp animationChannel.cpp animation.cpp animatable.cpp keyframeChannel.cpp driver.cpp)
 set(ARMATURE skeleton.cpp bone.cpp ikSolver.cpp)
@@ -121,8 +121,8 @@ if(BUILD_SAMPLES)
 	add_executable(particleEmitterSample particleEmitterSample.cpp)
 	target_link_libraries(particleEmitterSample ${LIB_NAME})
 
-	add_executable(raySample raySample.cpp)
-	target_link_libraries(raySample ${LIB_NAME})
+	add_executable(rayCasterSample rayCasterSample.cpp)
+	target_link_libraries(rayCasterSample ${LIB_NAME})
 
 	add_executable(pbrSample pbrSample.cpp)
 	target_link_libraries(pbrSample ${LIB_NAME})
